Add ClassLoader to InlinerOptions
Simplify how class loading is done during the tool execution, so it is
less error-prone, and should generally work better.
It is necessary for proper array construction, because we need to find
the component type to create an appropriate array.

Also handle all array creations, not only the ones configured in
BaseConfig.